Add testing framework to the project along with rules to compile test files.
tests.c and tests.h is center for all other tests files that runs all the tests.
Unit testing to libraries allow `perfection over abstraction` philosophy of the project. Seatest testing framework for c is incorporated into Class-DB because it is simple and independent library and do not club with the project. Testing is completely independent of the development code.